Added screen info, added initial start of abstract class for all animals, added config for hunger, changed hunger to a float

+++ src/main/java/mokiyoki/enhancedanimals/entity/EnhancedAnimal.java
@@ -4,9 +4,9 @@ import net.minecraft.inventory.Inventory;
 
 public interface EnhancedAnimal {
 
-    int getHunger();
+    float getHunger();
 
-    void decreaseHunger(int decreaseAmount);
+    void decreaseHunger(float decreaseAmount);
 
     Boolean isAnimalSleeping();
